requirements.txt
tkinter
Pillow
matplotlib
pix2text